Refactor the linter into a submodule.
